+++ src/jakson/stdinc.h
@@ -150,9 +150,6 @@ MAYBE_UNUSED static const char *basic_type_to_system_type_str(enum archive_field
         }
 }
 
-#define __CPP_VA_ARGS_11(a1, a2, a3, a4, a5, a6, a7, a8, a9, a10, a11, ...) a11
-#define VA_ARGS_LENGTH(...) __CPP_VA_ARGS_11(_, ## __VA_ARGS__, 9, 8, 7, 6, 5, 4, 3, 2, 1, 0)
-
 #if !defined(LOG_TRACE) || defined(NDEBUG)
 #define TRACE(tag, msg, ...) { }
 #else